Programa que ler nome e média de um aluno e guarda a situação dele em um dicionario python

+++ Exercicio 90.py
@@ -2,7 +2,7 @@
 alunos = []
 cadastro = dict()
 cadastro['Nome'] = str(input('Nome :'))
-cadastro['Media'] = int(input(f'Digite a média de {cadastro["Nome"]} :'))
+cadastro['Media'] = float(input(f'Digite a média de {cadastro["Nome"]} :'))
 print('-' * 30)
 print(f'{"Dados do aluno":^30}')
 print('-' * 30)
@@ -16,4 +16,3 @@ if alunos[0]["Media"] < 6:
 else:
     alunos[0]["situação"] = 'Aprovado'
     print(f'Situação : {alunos[0]["situação"]}')
-
